directory-service/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
python-logging-loki
requests
httpx

